+++ editors/blocknote-headless/src/vue/blocks/ImageSearchSelector.vue
